Fix account balance
