+++ core-rust/mesh-api-server/src/mesh_api/conversions/mod.rs
@@ -1,5 +1,6 @@
 mod addressing;
 mod block;
+mod common;
 mod context;
 mod errors;
 mod hashes;
@@ -7,6 +8,7 @@ mod numerics;
 
 pub use addressing::*;
 pub use block::*;
+pub use common::*;
 pub use context::*;
 pub use errors::*;
 pub use hashes::*;
